+++ src/components/map/sectors/MapSector.vue
@@ -9,7 +9,7 @@
         :settings="{
             position: [fir.lon, fir.lat],
             positioning: 'center-center',
-            stopEvent: isHovered,
+            stopEvent: isHovered && hasScroll,
         }"
         :z-index="15"
         @mouseleave="isHovered = false"
@@ -37,6 +37,7 @@
             #popup
         >
             <common-controller-info
+                ref="controllerInfo"
                 absolute
                 :controllers="[...locals.map(x => x.controller!), ...globals.map(x => x.controller!)]"
                 show-atis
@@ -60,6 +61,7 @@ import { useMapStore } from '~/store/map';
 import CommonControllerInfo from '~/components/common/vatsim/CommonControllerInfo.vue';
 import MapOverlay from '~/components/map/MapOverlay.vue';
 import { getAirportCountry } from '~/composables/airport';
+import { useScrollExists } from '~/composables';
 
 const props = defineProps({
     fir: {
@@ -73,6 +75,12 @@ const props = defineProps({
 });
 
 defineSlots<{ default: () => any }>();
+
+const controllerInfo = ref<{ $el: HTMLDivElement } | null>(null);
+
+const hasScroll = useScrollExists(computed(() => {
+    return controllerInfo.value?.$el.querySelector('.atc-popup_list');
+}));
 
 const mapStore = useMapStore();
 const vectorSource = inject<ShallowRef<VectorSource | null>>('vector-source')!;